[react] Upgrade Server Actions to canary
