+++ types/react/ts5.0/test/canary.tsx
@@ -1,5 +1,11 @@
 /// <reference types="../canary"/>
 
+// NOTE: forward declarations for tests
+declare var console: Console;
+interface Console {
+    log(...args: any[]): void;
+}
+
 const contextUsers = React.createContext(["HAL"]);
 const promisedUsers = Promise.resolve(["Dave"]);
 
@@ -84,4 +90,92 @@ function useCacheTest() {
         // @ts-expect-error -- experimental only
         refresh(() => "refresh");
     }
+}
+
+function useAsyncAction() {
+    const [isPending, startTransition] = React.useTransition();
+
+    function handleClick() {
+        // $ExpectType void
+        startTransition(async () => {});
+    }
+}
+
+function formActionsTest() {
+    <form
+        action={formData => {
+            // $ExpectType FormData
+            formData;
+        }}
+    >
+        <input type="text" name="title" defaultValue="Hello" />
+        <input
+            type="submit"
+            formAction={formData => {
+                // $ExpectType FormData
+                formData;
+            }}
+            value="Save"
+        />
+        <button
+            formAction={formData => {
+                // $ExpectType FormData
+                formData;
+            }}
+        >
+            Delete
+        </button>
+    </form>;
+}
+
+const useOptimistic = React.useOptimistic;
+function Optimistic() {
+    const savedCartSize = 0;
+    const [optimisticCartSize, addToOptimisticCart] = useOptimistic(savedCartSize, (prevSize, newItem) => {
+        // This is the default type for un-inferrable generics in TypeScript.
+        // To have a concrete type either type the second parameter in the reducer (see addToOptimisticCartTyped)
+        // or declare the type of the generic (see addToOptimisticCartTyped2)
+        // $ExpectType unknown
+        newItem;
+        console.log("Increment optimistic cart size for " + newItem);
+        return prevSize + 1;
+    });
+    // $ExpectType number
+    optimisticCartSize;
+
+    const [, addToOptimisticCartTyped] = useOptimistic(savedCartSize, (prevSize, newItem: string) => {
+        // $ExpectType string
+        newItem;
+        console.log("Increment optimistic cart size for " + newItem);
+        return prevSize + 1;
+    });
+    const [, addToOptimisticCartTyped2] = useOptimistic<number, string>(savedCartSize, (prevSize, newItem) => {
+        // $ExpectType string
+        newItem;
+        console.log("Increment optimistic cart size for " + newItem);
+        return prevSize + 1;
+    });
+
+    const addItemToCart = (item: unknown) => {
+        addToOptimisticCart(item);
+        addToOptimisticCartTyped(
+            // @ts-expect-error unknown is not assignable to string
+            item,
+        );
+        addToOptimisticCartTyped(String(item));
+        addToOptimisticCartTyped2(
+            // @ts-expect-error unknown is not assignable to string
+            item,
+        );
+        addToOptimisticCartTyped2(String(item));
+    };
+
+    const [state, setStateDefaultAction] = useOptimistic(1);
+    const handleClick = () => {
+        setStateDefaultAction(2);
+        setStateDefaultAction(() => 3);
+        setStateDefaultAction(n => n + 1);
+        // @ts-expect-error string is not assignable to number
+        setStateDefaultAction("4");
+    };
 }
